SSR/CSR Map fixed. should build now (#30)
